Fix build errors from upstream dependency changes
- anthropic.rs: pass &str to model() for Into<Model> compatibility
- openai.rs: migrate deprecated model_id() to model(ChatModel)
- export.rs: remove nonexistent summary_ja field, fix dataset -> datasets
- models.rs: widen PaperSection.index from i8 to i16 to match rsrpp

+++ src/models.rs
@@ -214,7 +214,7 @@ impl std::fmt::Display for SectionImportance {
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct PaperSection {
     /// Section index (order in paper)
-    pub index: i8,
+    pub index: i16,
 
     /// Section title (e.g., "Abstract", "Introduction")
     pub title: String,
